Refactor: 스타일드 분리
-colorimageselector와 createpaper 페이지를 제외한 나머지 파일 스타일드 분리
