+++ src/components/CreatePaper/BackgroundBox.jsx
@@ -11,3 +11,5 @@ export default function BackgroundBox() {
     </div>
   );
 }
+
+// 이 컴포넌트가 필요한지 고민중
