+++ src/router/router.jsx
@@ -1,5 +1,5 @@
 import { createBrowserRouter } from 'react-router-dom';
-import CreatePaper from '../components/CreatePaper/CreatePaper';
+import CreatePaper from '../pages/CreatePaper/CreatePaper';
 import Layout from '../components/Layout';
 import Home from '../pages/Home/Home';
 import List from '../pages/List/List';
